+++ app/src/main/AndroidManifest.xml
@@ -9,8 +9,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".FilmInfoActivity"/>
-        <activity android:name=".FilmsListActivity">
+        <activity android:name=".activities.FilmInfoActivity"/>
+        <activity android:name=".activities.FilmsListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
