+++ app/src/main/java/com/vkochenkov/filmscatalog/for_recycler/FilmViewHolder.kt
@@ -1,10 +1,11 @@
-package com.vkochenkov.filmscatalog
+package com.vkochenkov.filmscatalog.for_recycler
 
 import android.view.View
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.vkochenkov.filmscatalog.R
 import com.vkochenkov.filmscatalog.model.Film
 
 class FilmViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
