+++ app/src/main/res/layout/activity_film_info.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FilmInfoActivity">
+    tools:context=".activities.FilmInfoActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar_film_info"
